+++ dash/app/src/main/java/com/example/dlscj/dash/Dash.java
@@ -301,9 +301,9 @@ public class Dash extends Application {
 
     ///////////pattern/////////
     public native void initPatternMatch();
-    public native void initParam2Img(int time);
+    public native void initParam2Img(long time);
     public native void initNN(String path);
-    public native void updateParam2Img(int time, float v, float a);
+    public native void updateParam2Img(long time, float v, float a);
     public native void getPredicted(String predAlgorithm, float[] confs);
     public native boolean isValidPattern(int idx, float threshold);
     public native void isTop3(int idx, float threshold);
